Project is finished with no issues

+++ database/migrations/2024_10_26_215809_create_cart_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->uuid('id')->primary();
             $table->foreignUuid('cart_id')->nullable()->constrained('carts')->cascadeOnDelete();
             $table->foreignUuid('tour_id')->nullable()->constrained('tours')->cascadeOnDelete();
-            $table->integer('qty')->nullable();
+            $table->unsignedBigInteger('qty')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
